--- src/ru/MyOnlineShop/model/exeption/InvalidItemProduct.java
@@ -1,12 +1,0 @@
-package ru.MyOnlineShop.model.exeption;
-
-public class InvalidItemProduct extends Exception {
-
-
-    public InvalidItemProduct(String message) {
-        super(message);
-
-    }
-
-
-}
