--- src/ru/MyOnlineShop/model/product/food/bakery/Bread.java
@@ -1,17 +1,0 @@
-package ru.MyOnlineShop.model.product.food.bakery;
-
-import ru.MyOnlineShop.model.product.Product;
-
-import java.util.Collection;
-import java.util.Set;
-
-public class Bread extends Bakery {
-    public Bread(String categoryProduct, String typeProduct, String nameProduct, double price, int item, double expirationDate) {
-        super(categoryProduct, typeProduct, nameProduct, price, item, expirationDate);
-    }
-
-    public Bread(){
-
-    }
-
-}
