--- src/ru/MyOnlineShop/model/product/food/dairyProduct/DairyProduct.java
@@ -1,17 +1,0 @@
-package ru.MyOnlineShop.model.product.food.dairyProduct;
-import ru.MyOnlineShop.model.product.Product;
-import ru.MyOnlineShop.model.product.food.Food;
-import java.util.ArrayList;
-import java.util.List;
-
-public abstract class DairyProduct extends Food {
-    public DairyProduct(String categoryProduct, String typeProduct, String nameProduct, double price, int item, double expirationDate) {
-        super(categoryProduct, typeProduct, nameProduct, price, item, expirationDate);
-    }
-
-    public DairyProduct() {
-    }
-
-    public static List<Product> dairyProductList = new ArrayList<>();
-
-}
