--- src/ru/MyOnlineShop/model/product/nonFood/NonFood.java
@@ -1,20 +1,0 @@
-package ru.MyOnlineShop.model.product.nonFood;
-
-import ru.MyOnlineShop.model.product.Product;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public abstract class NonFood extends Product {
-
-    public NonFood(String categoryProduct, String typeProduct, String nameProduct, double price, int item) {
-        super(categoryProduct, typeProduct, nameProduct, price, item);
-    }
-
-    public NonFood(){
-
-    }
-
-    public static List<Product> nonFood = new ArrayList<>();
-
-}
