--- src/ru/MyOnlineShop/model/service/buyService/Delivery.java
@@ -1,4 +1,0 @@
-package ru.MyOnlineShop.model.service.buyService;
-
-public class Delivery {
-}
