+++ src/ru/myOnlineShop/model/exeption/InvalidCategoryProduct.java
@@ -1,0 +1,13 @@
+package ru.myOnlineShop.model.exeption;
+
+public class InvalidCategoryProduct extends Exception{
+    private String typeProduct;
+    public InvalidCategoryProduct(String message, String typeProduct){
+        super(message);
+        this.typeProduct = typeProduct;
+    }
+
+    public String getTypeProduct() {
+        return typeProduct;
+    }
+}
