+++ src/ru/myOnlineShop/model/exeption/InvalidItemProduct.java
@@ -1,0 +1,12 @@
+package ru.myOnlineShop.model.exeption;
+
+public class InvalidItemProduct extends Exception {
+
+
+    public InvalidItemProduct(String message) {
+        super(message);
+
+    }
+
+
+}
