+++ src/ru/myOnlineShop/model/product/food/bakery/Cracker.java
@@ -1,0 +1,12 @@
+package ru.myOnlineShop.model.product.food.bakery;
+
+public class Cracker extends Bakery {
+    public Cracker(String categoryProduct, String typeProduct, String nameProduct, double price, int item, double expirationDate) {
+        super(categoryProduct, typeProduct, nameProduct, price, item, expirationDate);
+    }
+
+    public Cracker() {
+
+    }
+
+}
