+++ src/ru/myOnlineShop/model/product/food/dairyProduct/Cheese.java
@@ -1,0 +1,12 @@
+package ru.myOnlineShop.model.product.food.dairyProduct;
+
+public class Cheese extends DairyProduct {
+    public Cheese(String categoryProduct, String typeProduct, String nameProduct, double price, int item, double expirationDate) {
+        super(categoryProduct, typeProduct, nameProduct, price, item, expirationDate);
+    }
+
+    public Cheese() {
+    }
+
+
+}
