+++ src/ru/myOnlineShop/model/service/buyService/Delivery.java
@@ -1,0 +1,4 @@
+package ru.myOnlineShop.model.service.buyService;
+
+public class Delivery {
+}
